+++ shopware/Component/TranslationImporter/ImportCSVCommand.php
@@ -63,6 +63,10 @@ final class ImportCSVCommand extends Command
         $domDocument = new \DOMDocument();
         $domDocument->loadXML(file_get_contents($pathToConfigXml));
         $row = fgetcsv($stream, null, ';'); // skip header
+
+        $result = $this->appender->append($domDocument, 'card.logger.title', 'Logger', $localeCode);
+        $output->writeln('<' . $result->getStatus() . '>' . $result->getMessage() . '</' . $result->getStatus() . '>');
+
         while ($row = fgetcsv($stream, null, ';')) {
             $key = $row[0];
             $key = $this->keyMappings[$key] ?? $key;
@@ -70,6 +74,7 @@ final class ImportCSVCommand extends Command
             $result = $this->appender->append($domDocument, $key, $text, $localeCode);
             $output->writeln('<' . $result->getStatus() . '>' . $result->getMessage() . '</' . $result->getStatus() . '>');
         }
+
         $domDocument->save($pathToConfigXml);
 
         return Command::SUCCESS;